pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "hatchling"]
build-backend = "hatchling.build"

[project]
name = "BLOOMDemo"
classifiers = ["Private :: Do Not Upload"]
version = "0"
requires-python = ">=3.10"
dependencies = [
    "ctranslate2",
    "transformers",
    "tensorflow",
    "Flask",
    "gunicorn",
    "ecs-logging",
    "pyyaml",
]
